api/requirements.txt
auto_mix_prep==0.2.0
fastapi==0.112.0
Pillow==9.3.0
Pillow==10.4.0
torch==2.3.1
torchvision==0.18.1
